// mobile/src/gen/types/ClientIn.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

/**
 * ClientIn
 */
export type ClientIn = {
  /**
   * @type string
   */
  name: string;
  /**
   * @default ""
   * @type string | undefined
   */
  phone?: string;
  /**
   * @default ""
   * @type string | undefined
   */
  email?: string;
  /**
   * @default ""
   * @type string | undefined
   */
  address?: string;
  /**
   * @default ""
   * @type string | undefined
   */
  notes?: string;
  rate: number | string;
  /**
   * @default "0"
   */
  cost?: number | string;
  /**
   * @default true
   * @type boolean | undefined
   */
  is_active?: boolean;
};